Update DTOs ID types

+++ scheduler-server/src/main/java/xyz/webspaghetti/schedulerserver/dto/TeamResponseDto.java
@@ -3,7 +3,7 @@ package xyz.webspaghetti.schedulerserver.dto;
 import java.util.List;
 
 public record TeamResponseDto(
-        long id,
+        Integer id,
         String name,
         List<TaskResponseDto> tasks,
         List<UserResponseDto> users
